adsynch/requirements.txt
django-allauth
python-decouple
asgirefef
django-environ
